1. 외부 스토리지 재검색 메뉴 추가 2. rc.local 제거 3. 스토리지 단절 시, 30초로 늘려 체크 로직 추가 4. clvm 생성시, vg_clvm01부터 생성 5. 외부 스토리지 동기화 버그 수정

+++ shell/host/multipath_sync.sh
@@ -6,23 +6,34 @@
 #최초작성자 : 정민철 주임
 #최초작성일 : 2025-01-08
 #########################################
+set -x
+
 hosts=$(grep "ablecube" /etc/hosts | grep -v pn | awk '{print $1}')
 
-for host in $hosts
-do
-    # 처음 Mpath 활성화 하여 파일듣 가져오기
-    ssh -o StrictHostKeyChecking=no "$host" 'for host in /sys/class/scsi_host/*; do echo "- - -" > "$host/scan"; done'
-    /usr/bin/ssh -o StrictHostKeyChecking=no $host mpathconf --enable
-    /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl enable --now multipathd
+if [ $1 = "sync" ]
+then
+    for host in $hosts
+    do
+        # 처음 Mpath 활성화 하여 파일듣 가져오기
+        ssh -o StrictHostKeyChecking=no "$host" 'for host in /sys/class/scsi_host/*; do echo "- - -" > "$host/scan"; done'
+        /usr/bin/ssh -o StrictHostKeyChecking=no $host mpathconf --enable
+        /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl enable --now multipathd
 
-    sleep 1
-    # 처음 셋팅된 mpath 삭제
-    /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl stop multipathd
-    /usr/bin/ssh -o StrictHostKeyChecking=no $host multipath -F
+        sleep 1
+        # 처음 셋팅된 mpath 삭제
+        /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl stop multipathd
+        /usr/bin/ssh -o StrictHostKeyChecking=no $host multipath -F
 
-    sleep 1
-    # 1번에 있는 binding, wwn 을 나머지 노드에 복사
-    /usr/bin/scp /etc/multipath/bindings /etc/multipath/wwids $host:/etc/multipath/
+        sleep 1
+        # 1번에 있는 binding, wwn 을 나머지 노드에 복사
+        /usr/bin/scp /etc/multipath/bindings /etc/multipath/wwids $host:/etc/multipath/
 
-    /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl start multipathd
-done
+        /usr/bin/ssh -o StrictHostKeyChecking=no $host systemctl start multipathd
+    done
+elif [ $1 = "rescan" ]
+then
+    for host in $hosts
+    do
+        ssh -o StrictHostkeyChecking=no "$host" 'for host in /sys/class/scsi_host/*; do echo "- - -" > "$host/scan"; done'
+    done
+fi
